requirements.txt
asgiref==3.4.1
click==8.0.3
colorama==0.4.4
Django==3.2.9
Flask==2.0.2
image==1.5.33
itsdangerous==2.0.1
Jinja2==3.0.3
jsonpickle==2.0.0
MarkupSafe==2.0.1
MouseInfo==0.1.3
numpy==1.21.4
opencv-python==4.5.4.58
Pillow==8.4.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt5-stubs==5.15.2.0
PyRect==0.1.4
PyScreeze==0.1.28
pytweening==1.0.4
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
Werkzeug==2.0.2
